requirements.txt
aiogram>=3.0
SQLAlchemy>=1.4
asyncpg>=0.27.0
python-dotenv
pydantic-settings
aiohttp
typing_extensions
